restructured GUI (gui attempt) into a package, and imported into pananal
