+++ GUI/validateData.py
@@ -1,5 +1,5 @@
-import FireCheck as fc
-from cellFix import getCellFix
+import GUI.FireCheck as fc
+from GUI.cellFix import getCellFix
 
 
 def checkFile(df):
