+++ pananal.py
@@ -1,0 +1,3 @@
+import GUI.gui as gp
+
+gp.run()
